Refactor View and ViewModel
I moved the slap and flip code from the View Slap function to their own
functions in the ViewModel onFlip and onSlap. I changed the board so
player slap returns whether or not the slap was valid rather than if the
game was won. I added a function to the board to check for a winner.